backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27

# Background Tasks
celery==5.3.4
redis==5.0.1

# ML/Classification
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
httpx==0.25.2